services/bot/requirements.txt
python-telegram-bot[job-queue]==21.3
asyncpg==0.29.0
redis[hiredis]>=5.0.0
anthropic>=0.49.0
google-genai>=1.10.0
python-dotenv==1.0.1
pytz==2024.1
reportlab==4.2.5
openpyxl==3.1.5
python-docx==1.1.2
httpx>=0.28.1
tenacity>=8.2.0
pydub>=0.25.1
Pillow>=10.0.0
pytest>=7.4.0
thefuzz>=0.22.1
python-Levenshtein>=0.25.1
